// sidecar/foundry-sidecar-client/src/interfaces/pf2e/actor/modifiers.d.ts
declare const MODIFIER_TYPES: Set<"untyped" | "item" | "ability" | "circumstance" | "potency" | "proficiency" | "status">;
type ModifierType = SetElement<typeof MODIFIER_TYPES>;
interface BaseRawModifier {
  /** An identifier for this modifier; should generally be a localization key (see en.json). */
  slug?: string;
  /** The display name of this modifier; can be a localization key (see en.json). */
  label: string;
  /** The actual numeric benefit/penalty that this modifier provides. */
  modifier?: number;
  /** The type of this modifier - modifiers of the same type do not stack (except for `untyped` modifiers). */
  type?: ModifierType;
  /** If the type is "ability", this should be set to a particular ability */
  ability?: AbilityString | null;
  /** Numeric adjustments to apply */
  adjustments?: ModifierAdjustment[];
  /** If true, this modifier will be applied to the final roll; if false, it will be ignored. */
  enabled?: boolean;
  /** If true, these custom dice are being ignored in the damage calculation. */
  ignored?: boolean;
  /** The source from which this modifier originates, if any. */
  source?: string | null;
  /** If true, this modifier is a custom player-provided modifier. */
  custom?: boolean;
  /** The damage type that this modifier does, if it modifies a damage roll. */
  damageType?: DamageType | null;
  /** The damage category */
  damageCategory?: DamageCategoryUnique | null;
  /** A predicate which determines when this modifier is active. */
  predicate?: RawPredicate;
  /** If true, this modifier is only active on a critical hit. */
  critical?: boolean | null;
  /** Any notes about this modifier. */
  notes?: string;
  /** The list of traits that this modifier gives to the underlying attack, if any. */
  traits?: string[];
  /** Hide this modifier in UIs if it is disabled */
  hideIfDisabled?: boolean;
}
interface ModifierAdjustment {
  /** A slug for matching against modifiers: `null` will match against all modifiers within a selector */
  slug: string | null;
  test: (options: string[] | Set<string>) => boolean;
  damageType?: DamageType;
  relabel?: string;
  suppress?: boolean;
  getNewValue?: (current: number) => number;
  getDamageType?: (current: DamageType | null) => DamageType | null;
}
interface RawModifier extends BaseRawModifier {
  modifier: number;
  /** Whether to use this bonus/penalty/modifier even if it isn't the greatest magnitude */
  force?: boolean;
}
// ...
declare class ModifierPF2e implements RawModifier {
  #private;
  slug: string;
  label: string;
  /** The value of the modifier */
  modifier: number;
  type: ModifierType;
  ability: AbilityString | null;
  adjustments: ModifierAdjustment[];
  force: boolean;
  enabled: boolean;
  ignored: boolean;
  /** The originating rule element of this modifier, if any: used to retrieve "parent" item roll options */
  rule: RuleElementPF2e | null;
  source: string | null;
  custom: boolean;
  damageType: DamageType | null;
  damageCategory: DamageCategoryUnique | null;
  predicate: PredicatePF2e;
  critical: boolean | null;
  traits: string[];
  notes: string;
  hideIfDisabled: boolean;
  /**
   * The "category" of modifier (a misnomer since bonuses and penalties aren't modifiers):
   * Recorded before adjustments in case of adjustment to zero
   */
  kind: "bonus" | "penalty" | "modifier";
  /**
   * Create a new modifier.
   * Legacy parameters:
   * @param name The name for the modifier; should generally be a localization key.
   * @param modifier The actual numeric benefit/penalty that this modifier provides.
   * @param type The type of the modifier - modifiers of the same type do not stack (except for `untyped` modifiers).
   * @param enabled If true, this modifier will be applied to the result; otherwise, it will not.
   * @param source The source from which this modifier originates, if any.
   * @param notes Any notes about this modifier.
   */
  constructor(args: ModifierObjectParams);
  constructor(...args: ModifierOrderedParams);
  get category(): this["damageCategory"];
  get value(): number;
  get signedValue(): string;
  /** Return a copy of this ModifierPF2e instance */
  clone(options?: { test?: Set<string> | string[] }): ModifierPF2e;
  /**
   * Get roll options for this modifier. The current data structure makes for occasional inability to distinguish
   * bonuses and penalties.
   */
  getRollOptions(): Set<string>;
  /** Sets the ignored property after testing the predicate */
  test(options: string[] | Set<string>): void;
  toObject(): Required<RawModifier>;
  toString(): string;
}
interface ModifierObjectParams extends RawModifier {
  name?: string;
  rule?: RuleElementPF2e | null;
}
type ModifierOrderedParams = [slug: string, modifier: number, type?: ModifierType, enabled?: boolean, ignored?: boolean, source?: string, notes?: string];